requirements.txt
# Dev environment
pip
autopep8
discord.py

# App
flask